backend/requirements.txt
fastapi[standard]
sqlmodel
uvicorn[standard]
asyncpg
pydantic-settings
alembic
httpx
argon2-cffi
